add VirtualBox dependency to install
`$ vagrant up` failed for me without having VirtualBox installed:

The provider 'virtualbox' that was requested to back the machine
'webserver' is reporting that it isn't usable on this system. The
reason is shown below:

Vagrant could not detect VirtualBox! Make sure VirtualBox is properly installed.
Vagrant uses the `VBoxManage` binary that ships with VirtualBox, and requires
this to be available on the PATH. If VirtualBox is installed, please find the
`VBoxManage` binary and add it to the PATH environmental variable.